remove old components if it's needed v2

+++ prow/scripts/cluster-integration/kyma-upgrade-k3d.sh
@@ -30,6 +30,7 @@ fi
 export TEST_INFRA_SOURCES_DIR="${KYMA_PROJECT_DIR}/test-infra"
 export KYMA_SOURCES_DIR="${KYMA_PROJECT_DIR}/kyma"
 export TEST_INFRA_CLUSTER_INTEGRATION_SCRIPTS="${TEST_INFRA_SOURCES_DIR}/prow/scripts/cluster-integration/helpers"
+export REMOVE_OLD_COMPONENTS="false"
 
 # shellcheck source=prow/scripts/lib/log.sh
 source "${TEST_INFRA_SOURCES_DIR}/prow/scripts/lib/log.sh"
@@ -99,6 +100,9 @@ popd
 log::success "Tests completed"
 
 if [[ $KYMA_MAJOR_VERSION == "1" ]]; then
+    # Extend scenario
+    REMOVE_OLD_COMPONENTS="true"
+
     # Upgrade kyma to latest 2.x release
     export KYMA_MAJOR_VERSION="2"
 
@@ -143,25 +147,27 @@ popd
 
 log::success "Tests completed"
 
-log::info "### Remove old components"
-helm delete core -n kyma-system
-helm delete console -n kyma-system
-helm delete dex -n kyma-system
-helm delete apiserver-proxy -n kyma-system
-helm delete iam-kubeconfig-service -n kyma-system
-helm delete testing -n kyma-system
-helm delete xip-patch -n kyma-installer
-helm delete permission-controller -n kyma-system
+if [[ ${REMOVE_OLD_COMPONENTS}=="true" ]]; then
+    log::info "### Remove old components"
+    helm delete core -n kyma-system
+    helm delete console -n kyma-system
+    helm delete dex -n kyma-system
+    helm delete apiserver-proxy -n kyma-system
+    helm delete iam-kubeconfig-service -n kyma-system
+    helm delete testing -n kyma-system
+    helm delete xip-patch -n kyma-installer
+    helm delete permission-controller -n kyma-system
 
-kubectl delete ns kyma-installer --ignore-not-found=true
+    kubectl delete ns kyma-installer --ignore-not-found=true
 
-log::info "### Run post-upgrade tests again to validate component removal"
+    log::info "### Run post-upgrade tests again to validate component removal"
 
-pushd "${KYMA_SOURCES_DIR}/tests/fast-integration"
-make ci-post-upgrade
-popd
+    pushd "${KYMA_SOURCES_DIR}/tests/fast-integration"
+    make ci-post-upgrade
+    popd
 
-log::success "Tests completed"
+    log::success "Tests completed"
+fi
 
 #!!! Must be at the end of the script !!!
 ERROR_LOGGING_GUARD="false"
